ai-agent/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.18.4
    # via ai-agent (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
    #   watchfiles
asyncpg==0.31.0
    # via ai-agent (pyproject.toml)
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via griffe
cryptography==46.0.5
    # via
    #   google-auth
    #   pyjwt
distro==1.9.0
    # via
    #   google-genai
    #   openai
fastapi==0.129.0
    # via ai-agent (pyproject.toml)
google-auth==2.48.0
    # via google-genai
google-genai==1.63.0
    # via ai-agent (pyproject.toml)
greenlet==3.3.1
    # via ai-agent (pyproject.toml)
griffe==1.15.0
    # via openai-agents
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   ai-agent (pyproject.toml)
    #   google-genai
    #   mcp
    #   openai
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.13.0
    # via openai
jsonschema==4.26.0
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
mcp==1.26.0
    # via
    #   ai-agent (pyproject.toml)
    #   openai-agents
openai==2.20.0
    # via openai-agents
openai-agents==0.8.4
    # via ai-agent (pyproject.toml)
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via
    #   ai-agent (pyproject.toml)
    #   mcp
pyjwt==2.11.0
    # via mcp
python-dotenv==1.2.1
    # via
    #   ai-agent (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via mcp
pyyaml==6.0.3
    # via uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.1.15
    # via tiktoken
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
    #   openai-agents
    #   tiktoken
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
sqlalchemy==2.0.46
    # via
    #   ai-agent (pyproject.toml)
    #   alembic
    #   sqlmodel
sqlmodel==0.0.33
    # via ai-agent (pyproject.toml)
sse-starlette==3.2.0
    # via mcp
starlette==0.52.1
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
tenacity==9.1.4
    # via google-genai
tiktoken==0.12.0
    # via ai-agent (pyproject.toml)
tqdm==4.67.3
    # via openai
types-requests==2.32.4.20260107
    # via openai-agents
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   google-genai
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   requests
    #   types-requests
uvicorn==0.40.0
    # via
    #   ai-agent (pyproject.toml)
    #   mcp
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
